Remove proposal endpoints
These endpoints now belong as extensions in the sequencer, if we are
going to keep them at all, since the proposer is part of the app-specific
header: it is not part of any built-in HotShot type and it is not needed
at all for the rest of the query service to work.

In fact, I'm not sure these endpoints have ever been useful for anything,
and we might just not add them back in the sequencer, thus deleting them
entirely.

This change also removes the `proposer_id` field from `BlockQueryData`
and `BlockSummaryQueryData`, as it is available as a field on the
`header` property of both those types.