Set block building parameters for local demo
Set a slightly longer block time and a much larger minimum number of
transactions. This gets the demo building blocks  with many transactions
(seems to be 10 or so on average with the settings of the load generator)
massively increasing the throughput. This fixes the
"transactions are not being finalized or being finalized too slowly"
warning I was seeing when running the demo.