Add test-node binary
This is a proof of concept for a sequencer test node binary. I think it
would make our lives much easier for testing integrations because we
will no longer have to maintain a separate docker compose project that
goes out of date all the time. I expect it would also be helpful for
other parties who would like to build on top of our sequencer.

Not sure if it makes sense to hide it behind the testing feature.
Currently it uses some code that is only intended for tests.

To try:

    cargo run --release --features testing --bin test-node

Submit some transactions

    cargo run --release --bin submit-transactions -- http://localhost:33000