Add initial implementation of summary endpoints for SQL
In order to provide the desired behavior, the explorer api needs to be able to retrieve
a list of summaries from its storage layer.  This can be done fairly easily with a SQL
backed storage layer, but the other layers may have some difficulty.

This commit adds the initial implementations for retrieving the data for the summary
requests from the SQL layer.  These two implementations are likely to be the most
difficult to implement out of the rest as they contain the most complicated queries
from the bunch.