Replace some explorer types with associated types
Based on feedback provided by @jbearer It is less restrictive to provide types
that can be provided by the underlying implementation.  This allows for flexibility
when it comes to the representation of the of the underlying type.  It also means
that the HotShot Query Service doesn't need to be concerned with these
underlying types and how they serialize or not, only that they do.

This commit adds these type restrictions.  After reviewing with @jbearer these
types are likely not what is really desired as it codifies some assumptions being
made about what is stored and how many elements are stored within a
BlockHeader.  For now we will continue with these types and adjust them as to
what their final design entails.