Add an adversarial stress tester for the query service
The new nasty-client program runs a loop that generates a random
workload for the sequencer query service. The workload is intended
to be adversarial and do things that a realistic rollup client
wouldn't do. The current workload specifically reproduces some
behavior we've seen on Gibraltar, such as opening very old streams
and keeping multiple streams open at once.

It should be easy to extend the client with new actions if we want
to test additional behaviors.