Remove Explorer implementation for non sql storage
The Explorer endpoints only make sense for a hot shot query
service that is backed by an archival mode SQL source.  This
is due to the need to run relational queries when retrieving
the data that the Explorer endpoints return.

This removes the implementations of the ExplorerStorage for
the FileSystemStorage and the NoStorage.