Add filter parameters for get transaction summaries
The explorer endpoint for retrieving the transaction summaries needs
the ability to decode some optional filters.  Currently tide_disco does
not allow us to retrieve query parameters from a URL, so it's easier
to add these filters as optional path parameters.

This adds those filter path parameters as optional suffixes to the
normal request paths.  This will likely need to change in the future
if we can introduce and decode query parameters from the request
URL.

This also adds new SQL queries and parameters for retrieving
transaction summaries with the ability to consider an additional
filter.  The only filter supported at the moment is the block, as the
rollups query cannot really be performed unless we have a table
to join against, and populate that records the rollup namespace at
the transaction level.  For now, rollup filters return nothing.