Add some tests for explorer api
We need some tests for the explorer api in order to prove it's operating
correctly.  For now this only needs some basic checks, but the tests
will need modifications in the future in order to account for the edge
cases.

This adds some initial test cases for the explorer api to test that these
requests succeed without issue.