Add search endpoint for explorer
The block explorer needs to be able to perform a search for blocks,
transactions, and potentially more things in the future.  For now we
can limit it to transactions and blocks.  At the very least this search
term needs to be able to find blocks and transactions with a partial
match to their corresponding hash codes.

Add a basic implementation of the search functionality to retrieve
a small set of blocks and transactions should their transactions
match the prefix of a given search term.