Implement retry logic for nasty-client
The goal is only to emit an ERROR level log in the case of a real
problem that indicates a bug, so that we can easily use this tool
for alerts. In some cases, transient API failures are expected,
such as when an API server is missing data and has yet to fetch it
from a peer. In these cases, we will retry several times, emitting
only WARN level logs until we finally ERROR after exceeding the
maximum number of retries. If we retry and eventually succeed, we
perform a rigorous test to check that the operation is fully healed;
ie API endpoints are allowed to fail transiently but our expected to
be self-healing and not remain flaky for very long.