Handle WebSockets timeouts
The server is allowed to close idle connections after a long time.
The random nature of the client creates such idle connections. We
don't consider it an error if an idle connection fails, but will
automatically refresh the connection and try again. If a recently
used connection fails, it _is_ considered an error.