Add fields to TransactionSummary
In order to effectively target teh correct transaction for the Block
Explorer, we need to know the correct offset for the transaction.  In
addition it can also help to know the total number of transactions
within the block for that transaction.

Adds `offset` and `num_transaction`s to `TransactionSummary` for
reference.