Update Explorer API tests to check consistency
At the moment the tests for the explorer API just ensure that there is
a successful response.  This doesn't check the data coming back is
what is expected.

To address this the tests are modified to check the values being
returned within the values retrieved from the explorer summary. These
values should match.  In addition the response types have been
modified in order to support checks for equality.