Add comments for various Explorer types
In order to better document the intention behind the existence of
some of the new types introduced by the Explorer API.  This can
provide some better context as to the intention behind these types
and can be useful for future maintainers.

Add comments to some of these undocumented types and trait
methods to provide insight as to what their use and purpose is.