Add context to SQL query
The get_block_summaries function that targets blocks summaries
starting at the given block hash can be confusing as it seems odd to
target a block by a hash, to then return a result from the same table
with a different filter condition.

This adds a comment to motivate the reasoning behind this query
and why it is performed in two separate steps.