Adjust query branch flow
The get_block_summaries data stream is not as straight forward with
it's branch flow as it could be.  There's a map called on a result that
will get branched off of anyway.  As a result the map chains immediately
into another map.  This isn't as easy to follow as it could be and results
in two anonymous functions declared within each other.

Adjust the control flow to unwrap the Result for the stream, and then
use only a single map to transform the row from the stream.

Adjust queries to remove complication

Some of the queries for the Explorer are needlessly complicated when
it comes to their filtering criteria.  There are sub-queries and
aggregations being made that can be simplified into single statements
or clauses.

Adjust these queries to take a more simple representation.