Modify match statement logic
The match statement that is matching based on a tuple has too many
SQL query statements that are too similar.  This causes the subtle
variations between the statements to not stand out as much.  For
clarity these statements can be broken up, as much as possible, in
order to allow them to impart the logical differences as much as
possible.

Splits up the 9 potentially different queries into two different phases of
3 statements each.  This should preserve the previous logic, while
making the filtered portion more clear.