Remove ExplorerDataSource trait
The `ExplorerDataSource` trait seems to serve no real purpose other
than an extra layer of distinction.  Other data sources in the project
allow for other implementations of the various `DataSource`s such that
their data can be sourced from other means.  For the Explorer APIs its
only  supported `DataSource` is the `Storage` layer of SQL.  As such
this distinction doesn't make much sense, and only leads to confusion.

In order to correct this, we remove the `ExplorerDataSource` trait and
instead opt to move its definitions to `ExplorerStorage`.  This allows us
to keep the errors that have already been defined.

In addition we  remove a requirement on the `ExplorerStorage` trait
needing the `AvailabilityStorage` as this is an implementation detail
rather than a hard referenced requirement.

We also add more errors to cover the cases where these new errors
can be wrapped up in their relevant enums.  For convenience we also
implement `From` on several of these errors to make them work with
the `branch` operator.

To make everything work the same we implement `ExplorerStorage`
instead of `ExplorerDataSource` for the types previously had the now
removed trait.