Rename `explorer::data_source` to `explorer::query_data`
Since the `ExplorerDataSource` trait has been removed, it doesn't make
sense to keep a module named `data_source` around for  explorer.  Since the
remaining contents of `explorer::data_source` are essentially type
information for the results of queries, and there is already a president for
query types to be stored in `query_data`  modules, we should rename
`explorer::data_source` to `explorer::query_data`.