Submit transactions to builder private mempool
Add an option in `submit-transactions` to submit to a builder. Add
a second transaction generator task in the demos to submit to the
builder. This both serves as a test of the builder API and allows
us to directly compare the latencies of transactions submitted via
the public and private mempools.