Make file system persistence updates atomic
Write all updates to temp files, and only replace the main file
at the very end using an atomic rename. This fixes a bug observed
in staging where the saved anchor leaf file for a node was empty
after being restarted, likely because the node was shut down in the
middle of updating this file, after truncating it but before flushing
the new data.