Fail validation, but don't panic, if builder fee exceeds `u64::MAX`
Since the builder fee is actually being charged now, instead of silently
ignored, I had to update all the tests to prefund the builder account. I
deleted `test_catchup` because it wasn't working nicely with prefunding
the builder account with a forgotten state, but it wasn't testing anything
that the newer `test_catchup_live` wasn't also testing in a better way.