Adjust tests involving transactions for Explorer API
The tests for the Explorer API currently weren't being executed when
dealing with the transactions.  This is due to the Mock Network not
having any transactions being generated and submitted to the
network.

In addition, for the transaction tests themselves, we were not covering
all of the path edge-cases involving transaction summary retrieval.
As such, this adds more cases to cover everything except the rollup
filter, as that is not yet implemented.

These new tests highlighted some potential query logic errors that
would result in inconsistent retrieval behavior.