Replace `.unwrap` calls with `Err` results
In the `get_transaction_detail` code there are locations where the
specific transaction is attempted to be retrieved.  In these cases they
are explicitly unwrapped instead of recognizing that they might not
be included in the returned `transaction stream`.  In this case we
favor returning the `Err` variant of a `Result` instead of using an
`unwrap` call that could panic.