Set working RUSTFLAGS in .cargo/config.toml
Avoids developers running into compilation errors due to RUSTFLAGS.

Finally addresses #845

This will be used for all rustc invocations unless the RUSTFLAGS env var
is set. In that case the env var will override the value in the config
file.

Removes setting of `--cfg hotshot_example` (except for on the CI). This
does lead to some warnings when compiling the tests in with `--release`
but requires less configuration. I think the main problem here is that
we need to run tests with `--release` and we should instead fix that by
for example enabling optimization of dependencies and leave the warning
there until we fix it.