Use local database for catchup when possible
There is a potential race condition in state persistence/restartability.
Since state persistence generally happens asynchronously, it is possible
(likely even) that a node gets shut down after persisting a decided leaf
but before computing the corresponding state. Then, when the node restarts,
it is unable to load the state corresponding to the saved leaf, and must
reconstruct this state lazily as part of catchup. If this happens to all
nodes, then there is no node able to serve catchup queries, and the system
stalls.

We want to keep the asynchronous state catchup, because it is extremely
simple and resilient to all kinds of failures in the underlying event
processing, and because it's auxiliary in a sense: as long as we have
the block data, we have all the information we need to reconstruct the
state at any time, and so it makes no sense to make the state computation
blocking.

In light of this, one potential solution would just make the state loading
at startup wait/retry until the state reconstruction loop has reconstructed
the desired state. However, a better, more generalized solution is to incorporate
the state storage into the normal catchup flow, which already has a retry loop.
This removes the need to load entire state snapshots into memory at all, which
is very nice: it is a step towards having the in-memory `ValidatedState` being
just a cache for the full persisted state. This also further unifies the restart
procedure with the normal catchup procedure. In other words, it is as if a node
that has restarted has merely been offline for some amount of time.

Other changes:
* `init_node` takes persistence _options_ now, rather than an instantiated
  persistence object, so that we can construct multiple connections, one
  for catchup and one for event processing. This leads to a lot of changes
  that are mostly uninteresting.
* `node_id` moved into `InstanceState` so we have easy access to it for
  logging
* numerous improvements to logging/instrumenting