Remove "wait for state storage to be updated" loop in `test_restart`
This loop was there to avoid shutting down the network before the
merklized state was persisted, otherwise the network would not start
up again properly and the test would fail -- exactly the bug that this
change is designed to fix! As of now the test passes without this extra
wait, which is good confirmation that the change is working as intended.