Add explorer API to SQL backed images
The explorer API that is available in the hotshot query servicer is
available to archival nodes that are backed by postgres sql.  In order
to make the explorer API available to consume, it needs to be exposed
as a module.

Adds the explorer API so that it is available for all `storage-sql`
featured nodes.