Move fee contract address to chain config and make it optional
The collection of fees (or lack thereof) is not completely controlled
by ChainConfig. This makes it possible to launch without fees (even
without a fee contract) and upgrade to enable fees at a later date.
The only code change required to do this upgraded will be the ability
to change the ChainConfig by upgrading. After that, only a configuration
change is need to change the ChainConfig to set the fee contract and
base fee.