Add explorer module option
The behaviors, features, and modules available within the `sequencer`
binary are enabled via a command specified within the command line
arguments.  Since the explorer is another such module, it too should be
configured to be enabled or disabled based on the same configuration
pattern.

This adds the explorer configuration option for optionally enabling the
explorer module.  It also links the dependencies of the module in order
to ensure that is is enabled correctly.