Define a genesis config file
This aims to separate onchain data, which must be the same for all
nodes (chain ID, base fee, etc) from node-specific configuration.
Node configuration continues to be done via environment variables,
whereas onchain data is defined in a configuration file which we
can distribute to all node operators when launching a network. This
should greatly simplify coordination and reduce the frequency of
errors due to misconfigured nodes.