Remove namespace_ids_for_offset from `ExplorerHeader`
Add namespace_id to `ExplorerTransaction`

The trait `ExplorerHeader` had a requirement that the `NamespaceId`s
for a given transaction offset could be retrieved from the `Header` of
a block.  Considering implementation details, the Block Header doesn't
have the data required to make this determination.  The Transaction,
however, is capable of determining which `NamespaceId` it belongs to.

To correct this issue a new trait is created to restrict the `Transaction`
data type in order to be able to retrieve the `NamespaceId` from it.  In
addition the `namespace_ids_for_offset` is removed as it doesn't
make sense.