Add a comment explaining stub implementation
Add TODO mark for stub implementation to comment

Add Tracked issue reference to comment

The `reward_balance` method for the `ExplorerHeader` implementation
has a stub implementation as the Block Reward is not currently supported
within the sequencer.  In order to document this effectively, a comment
has been added to track it as such, along with a TODO that explains what
change will need to happen, and an issue for added visibility / lookup.