Add a wrapper `Auction` type w/ rudimentary state machine
I'm probably going off on a limb, but its easier for me to think about
the types if there is some logic tying them together. Ultimately I
think we will expose a function to an event handler that will make the
gears turn. I believe `ViewNumber` can be passed in, and I'm guessing
`BidTx` we be as well.