Fix explorer summary histogram data order
The histogram data returned by the `get_explorer_summary` endpoint is
returned in the wrong order.  The data is meant to be in the order of
ascending block height, but it is returned in descending block height
order.  This is due to the way we retrieve the last 50 blocks.

To fix this the query that retrieves the latest 50 blocks is changed slightly.
Using a sub-query, we can target the last 50 blocks, then re-sort the rows
based on the block height ascending to get the correct ordering.