Fix WebSocket timeout handling in nasty-client
The server might close an idle connection after 60s. This should not
be considered an error. Currently, though, it is causing errors, because
our caluclation of idle time is not correct. We are using time since last
read, but a read doesn't necessarily mean traffic over the connection. We
might be reading a message that was sent long ago and has been sitting in
our TCP buffer while the connection has been idle.

To more accurately gauge how long the connection has been idle, we use time
since the last _blocking_ read. Thus, we are less likely to produce an error,
and more likely to just chalk ConnectionReset errors up to a stale stream and
refresh the stream and continue.