Fix: TmpDb::init() wait for psql (#602)
Previously `TmpDb::init()` waited for postgres to become ready,
but the logic would panic if `psql` is not available on the
host. But, we should not have any expectation that `psql` be available
on the host. This PR updates the wait loop to run a query inside
the docker container.

---------

Co-authored-by: tbro <tbro@users.noreply.github.com>