Serialize L1 block numbers as hex strings
- Add newtype for L1 block number.
- Implement serialization in backwards compatible manner. Although there
  is currently no test for this backwards compatibility.
- Don't ignore "actual" files created from reference tests.
- Remove "actual" files for successful tests, if they exist.

After implementing this and looking at the JSON file I'm now of the
opinion that we should throw the compatibility with L1 away. It looks
rather messy and I don't see what we really gain from it. I now think we
should just serialize everything in the way we think is best, most
pretty or whatever.

Close #1507