Remove unnecessary write lock for submit endpoint
The submit endpoint took a write lock on the state by default, since
it is a POST endpoint. However, this lock is unnecessary, because
internally all we do is send an event, we don't mutate any state.
By using `at` instead of `post`, we can manually control the locking,
and take just a read lock. This should help significantly in avoiding
queues on the API state lock when many transactions are being submitted.