Do not hold API state lock while waiting for fetch
The default behavior of the availability API is to wait up to 500ms
for data to be fetched if it is missing. With the default tide-disco
locking, we are holding a shared lock on the state this entire time.
This can cause queues to form where a writer has to wait a while for
a bunch of readers to release their locks, and then subsequent readers
have to wait even longer for the writer to get the lock and release it.

This change implements manually locking so that we only old the lock
while _spawning_ the fetch, not while awaiting it.