Ensure termination of availability NoStorage test
This test has always been producing blocks faster than they can be
fetched (producing 1 block a second, waiting up to 5 seconds to fetch
a block). This should make the test take forever, since it never finds
the end of the ledger. The only reason it was terminating before is that
it was holding a lock on storage for the entire 5s it was waiting for
blocks to be fetched, so newly produced blocks were not inserted into
the database until after the request had failed.