Improve reliability of TmpDb startup
Because we were only checking for database readiness from inside
the container, there was a case where we would consider the TmpDb
"ready" and allow connections before the Docker daemon had actually
exposed the database port to the host. This change explicitly checks
that the port is exposed on the host before returning, with no extra
dependencies on the host side.