Add node-metrics crate
The `node-metrics` is designed to provide information about the block
chain and the nodes connected to the block chain network.  The name of
the crate was suggested by Charles so that it might be used by other
systems.  However, in general it is intended to be used to serve
information for the Node Validator Dashboard.

This crates defines various messages and logic for consuming and
distributing information about the Espresso Block Chain and the nodes
connected to the Espresso Sequencer network.

The current implementation is a basic implementation in an effort to get
the simplest setup working.  At the moment, there are two issues that
prevent anything from building.

- The first is that the `Sink` implementation of `Connection` from
  `tide-disco` requires the message being send to bea  borrowed value
  instead of the value itself.  I feel that this is unexpected, and it makes
  it difficult to actually send messages as a passthrough.

- The second is that the `sequencer` crate is being consumed in an effort
  to extract the `SeqTypes` type.  By doing this, the idea is that we should
  be able to use the actual types defined by the sequencer so that we are
  able to consume the exact information we will be given in a production
  environment.  However, due to this import the tests that have been
  written cannot be evaluated.  I believe that this means that this cannot
  work as expected, and will need to be rethought.