Relocate the leaf stream construction
There will be a need to attempt to be resilient and be able to recover from certain
classes of errors. As such it is helpful to be able to create a Stream from a given
Client as needed.  This will allow for future changes to be able to re create the
stream in the event of a disconnection (which will happen).