Refactor `BitVec` for voters to be u16
By default `BitVec`s are `usize` in their representation.  This works great for systems that
are able to decode `u64` types from a JSON representation, but for languages that have a
unified number system, such as Javascript, these are not representable accurately when
decoded.  In order to support these cases it is easier to just have the `BitVec` represent it's
data in a format that is supported.

`u32` could be used here, but for maximum compatibility, I've chosen to use `u16` instead.