Replace ip addresses in node identity with public url
Due to the potential difficulty there can be in knowing your standing
IP Address on cloud platforms, the requirement of knowing the ip address
of a node has been replaced with a public url instead.  This public url
should be the base url for an API endpoint for the sequencer.