Add functions to parse Node Identity information from prometheus metrics
We intended to be able to retrieve Node Identity information from the prometheus
metrics.  As such we need a function that can handle this interaction. Using the
plugin for `prometheus-parse`, and `reqwest` we can retrieve the data given a
valid base URL to work with.  From there we can parse the data that is available
in the resulting `Scrape` object to fill in the missing pieces of Node Identity
information.