Update API database in sync with consensus storage
Previously, the database used by the query API was populated from a
completely separate event handling task than the consensus storage.
This could lead to a situation where consensus storage has already
been updated with a newly decided leaf, but API storage has not, and
then the node restarts, so that consensus things it is on a later leaf,
but the query API has never and will never see this leaf, and thus cannot
make it available: a DA failure.

With this change, the query database is now populated from the consensus
storage, so that consensus storage is authoritative, and the query datbase
is guaranteed to always eventually reflect the status of consensus storage.
The movement of data from consensus storage to query storage is tied in with
consensus garbage collection, so that we do not delete any data until we are
sure it has been recorded in the DA database, if appropriate.

This also obsoletes the in-memory payload storage in HotShot, since we are
now able to load full payloads from storage on each decide, if available.