Allow nodes to fetch network config from peers (#1711)
This allows nodes to fetch the network config from the `config` API of
another node, instead of/in addition to local storage and the
orchestrator. This could be useful when a node in an existing network is
migrating to new storage, or when storage has been lost. In particular,
Kudasai is currently not running their Cappuccino nodes because (for
reasons still unknown) they lost their persisted network config, and now
have no way of getting it back since the orchestrator is already shut
down.

* Adds a `PublicNetworkConfig` which wraps a `PublicHotShotConfig`
* Has the `config/hotshot` endpoint return the network config instead of
just the HotShot config, which is strictly more information
* Adds functions to convert a `PublicNetworkConfig` to a proper
`NetworkConfig` based on a given set of private keys
* Adds configuration logic to fetch the network config from a peer on
startup

* Fix any issues with the network config being lost from persistent
storage

* New config serialization in `api/data_source.rs`
* Configuration logic in `lib.rs`