Fix test_process_client_handling_stream_subscribe_voteres test
The test `test_process_client_handling_stream_subscribe_voters` never completes and just
runs forever.  The reason for this is due to a flaw in the creation of the test itself.
The test is intended to check on the real-time submission of voters data, and the
 distribution of that data to the subscribers.  However, the setup has the users subscribe
to `NodeIdentity` updates instead of `Voters` updates.  To fix the issue we just need to
change the subscribe calls to refer to voters instead.