Add process_node_identity_stream
We need to process a stream of updating node identity information as it comes in.
In this way we can ensure that we have a single location where we update the local
stored state of node identity information, and the hand-off to potential subscribers
of the real time updates of the node identity information.