Add Node Identity information to Sequencer metrics reporting
In order for the sequencers to have reportable data available to the Node Validator, they need
to expose the identifying data to the node validator in a public way.  In this case, we want
to record the information in the status/metrics endpoint.  However, we would also like this
information to be configurable by the end user via Environment variables.

This change adds the labels to the various metrics areas, and updates the process-compose.yaml
and docker-compose.yaml files to configure some values for the sequencers for local testing.