Replace Sender with generic parameter
In order to make the implementation of the code more flexible and to support
different types of `Sink`s, the hard-coded use of `Sender` has been replaced
with a generic parameter `K` instead. This allows for us to swap out different
types of `Sink`s and `Stream`s throughtout the code base.