Fix inconsistent BLSPubKey decoding from str
There are two ways that the BLSPubKey is being decoded from the string that yield
inconsistent results. The first is `BLSPubKey::from_str` which seems to be the correct
way to decode these values.  The second was to decode the `str` into a `TaggedBase64`
and then to take the containted data and pass it to `BLSPubKey::from_bytes` which matches
the majority of the way, but seems to differ for some reason.