Improve performance issue in `create_block_detail_from_leaf`
There's a performance issue in `create_block_detail_from_leaf` concerning creating a
`BlockDetail` object from the `Leaf` and its `Payload.  The primary cause of this
performance issue involves attempting to determine the number of transactions and the
byte size of the `Payload`. For both of these, the data was retrieved from the trait
`QueryablePayload`.  `BlockPayload::num_transactions` ultimately calls a helper
method called `Queryable::transactions`.  `Queryable::transactions` is implemented by
`sequencer::block::full_payload::payload::Payload` by calling
`QueryablePaylaod::enumerate`.  The implementation of `QueryablePayload::enumerate`
ultimately invokes `QueryablePayload::transaction` which is implemented by calling
`QueryablePayload::transaction_with_proof`.  This means that these convenience
functions ultimately involve calculating a proof which does not get used.

To fix this, we rely on the `QueryablePayload::iter` instead, as it is able to
determine the number of transactions, and using it we can also take advantage of
`Payload::transaction` to retrieve the transaction from the iterated `Index` objects.
From there we can get the `size` of each `Trasnaction` contained within the `Payload`
without needing to invoke proof considerations.