Rename Node Identity Environment variables
The Node Identity environment variables don't match the existing name scheme that
we use for the other Sequencer environment variables.  Namely they don't start
with the prefix "ESPRESSO_SEQUENCE".  I believe this may cause a disconnect that
would imply that these variables are not used within the sequencer themselves, but
this is actually the only place where these variables are used.  As such, this
renames them to match the other variables.