Add optional processing of HotShot Events
In order to support the ability to process external messages, we need the ability to
process HotShot events.  In genreal, we only really need to know about RollCallResponse
messages, as that will help us to know when a node is responding to our RollCallRequest.

We also need the ability to send RollCallRequests so we can get the response, and wait
for any incoming new urls to scan.