Remove leaf retrieval from `create_node_validator_processing` call
With the ability to potentially process HotShot Events via the async task
`HotShotEventProcessingTask`, this means that we technically could have access to the `Leaf`
nodes via `Decide` events.  This can be used as an alternative to retrieving thee `Leaf`
from the HotShot Query Service as is done by the `ProcessProduceLeafStreamTask` async
task.  As such, it doesn't make sense to impose one over the other in the call to
`create_node_validator_processing` anymore.  So we remove it in order to allow the
consumer to make the call to which ever mode they may prefer instead.

This allows for flexibility when it comes to how we source this information.