Add run_standalone_service call
Add main

In order for the node-validator api to run as an independent service, it needs to have its
own main function for it to run.  This is implemented with the assumption that it can
connect to the CDN in order to publish and receive RollCall events.  Additionally, it is
assumped that some initial known node urls can be provided immediately to aid in initial
node data population, if necessary.  The leaves are also sourced from a hotshot query
service availability api.