Refactor separate CDN tasks, and make optional
The CDN processing tasks are define as an inline anonymous code block
/ function that is defined entirely within the
`run_standalone_service`. Additionally, these processes are not
optional.  Following discussions with @jbearer at the Front End Weekly
meeting, the CDN should probably not be utilized as it acts as a
hard-barrier that prevents third-parties from being able to build their
own dashboards using the data we are providing.

As such, the background tasks have been moved into their own types for
tracking, and the CDN configuration for them has been made entirely
optional.