Refactor simplify InternalClientMessage
The `InternalClientMessage` enum largely follows the `ClientMessage`, and as such it has many
duplicate cases.  As pointed out by @jbearer in this comment:
https://github.com/EspressoSystems/espresso-sequencer/pull/1771#discussion_r1700458919.

It makes more sense to wrap the `ClientMessage`s instead of having duplicate cases for them.
This also has the added benefit of cleaning up much of boilerplate implementation details.

Additionally, this relocates the `PartialEq` implementation into the test module so that it is
only defined for unit tests.