Refactor `DataState` initialization to avoid mutation
The `new` call for `DataState` takes a `stake_table` as part of its arguments.  Currently, the
`DataState` gets called with `Default::default()` arguments, and then the `stake_table` is
replaced within it.  This is unnecessary as one of the next steps of initialization is to
retrieve the `stake_table` from the `stake_table_url_base` url.

Instead this should just defer the creation of the `DataState` until after the `stake_table`
is retrieved.

Suggested by @jbearer in this discussion:
https://github.com/EspressoSystems/espresso-sequencer/pull/1771#discussion_r1700515508