Fix Identity value population
From discussion with @jbearer
https://github.com/EspressoSystems/espresso-sequencer/pull/1771#discussion_r1701015549

Based on the converesation with Jeb the default value can be utilized to populate the operating
system value from the environment directly.  While implementing this fix, it was also discovered
that the previous way of populating the Identity using `Default::default` actually prevented the
program from running at all due to bad initialization.

I haven't determined the exact reason, but by switching back to default_value population, we
can revert to using the `flatten` option, and the issue seems to be fixed as a result.