Remove wallet_address from NodeIdentity
Discussion with @jbearer:
https://github.com/EspressoSystems/espresso-sequencer/pull/1771#discussion_r1701001580

Right now we don't have a meaningful definition of what an address for a Sequencer is.  As a
result it makes more sense to omit the wallet address field rather than to keep it in an effort
to create a forward compatible definition.

This commit removes the definition, population, and representation of the wallet address from
both the sequencer, and from the node-validator api.

This will require a change in the Node Validator UI, as it expects a wallet_address to be
present.